requirements.txt
Django~=3.1.0
pytz~=2020.1
flake8~=3.8
djangorestframework~=3.12.0
Pillow~=8.0.1